Fix: Issue faulty URLs returned by BuildInfoResultAction
This commit solves:
  https://github.com/jfrog/jenkins-artifactory-plugin/issues/454

As per @jons suggestion, we stop escaping URLs when stored in the
BuildInfoResultAction, but instead _only_ when returned.

To ensure that old BuildInfos are handled correctly, a readResolve()
detects the string "%20%3A%3A%20", which indicates an already
escaped URL (while it is possible to create a Jenkins job name
that includes this String, it should in practice never happen).